+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### Modules
 * [Misc] Create Ubuntu specific copies of bcc modules to reflect the Debian package adding -bpfcc to the binary file names.
+* [New Module] Add enadiag net diagnostic module.
 
 #### Testing
 * [Bugfix] Update requirements_test.txt to lock versions of transient dependencies causing issues for py2.7 builds